+++ spec/gamefic/scriptable_spec.rb
@@ -177,18 +177,4 @@ describe Gamefic::Scriptable do
     plot = klass.new
     expect(plot.foo.name).to eq('foo')
   end
-
-  it 'makes no_scripts modules' do
-    original = Module.new do
-      extend Gamefic::Scriptable
-
-      def foo; end
-
-      script {}
-    end
-    no_scripts = original.no_scripts
-    expect(no_scripts.public_instance_methods).to eq(original.public_instance_methods)
-    expect(original.blocks).to be_one
-    expect(no_scripts.methods).not_to include(:blocks)
-  end
 end